move icons, remove unused color
