+++ src/home/TabHome.tsx
@@ -15,12 +15,12 @@ import Touchable from 'src/components/Touchable'
 import { ALERT_BANNER_DURATION, DEFAULT_TESTNET, SHOW_TESTNET_BANNER } from 'src/config'
 import { CICOFlow } from 'src/fiatExchanges/utils'
 import { refreshAllBalances, visitHome } from 'src/home/actions'
-import ArrowVertical from 'src/icons/ArrowVertical'
 import Add from 'src/icons/quick-actions/Add'
-import Send from 'src/icons/Send'
-import Swap from 'src/icons/Swap'
 import SwapArrows from 'src/icons/SwapArrows'
-import Withdraw from 'src/icons/Withdraw'
+import ArrowVertical from 'src/icons/tab-home/ArrowVertical'
+import Send from 'src/icons/tab-home/Send'
+import Swap from 'src/icons/tab-home/Swap'
+import Withdraw from 'src/icons/tab-home/Withdraw'
 import { importContacts } from 'src/identity/actions'
 import { navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
