+++ src/icons/tab-home/Withdraw.tsx
